Add Simulator state field and booting capability
