+++ src/Microsoft.DotNet.XHarness.iOS.Shared/Hardware/SimDevicePair.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+
+#nullable enable
+namespace Microsoft.DotNet.XHarness.iOS.Shared.Hardware;
+
+public class SimDevicePair
+{
+    public string UDID { get; }
+    public string Companion { get; }
+    public string Gizmo { get; }
+
+    public SimDevicePair(string uDID, string companion, string gizmo)
+    {
+        UDID = uDID;
+        Companion = companion;
+        Gizmo = gizmo;
+    }
+}
